+++ src/main/java/com/login/app/controlador/RegistroControlador.java
@@ -1,13 +1,25 @@
 package com.login.app.controlador;
 
+import com.login.app.servicio.UsuarioServicio;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class RegistroControlador {
 
+    @Autowired
+    private UsuarioServicio servicio;
+
     @GetMapping("/login")
     public String iniciarSesion(){
         return "login";
+    }
+
+    @GetMapping("/")
+    public String verPaginaDeInicio(Model modelo){
+        modelo.addAttribute("usuarios",servicio.listarUsuarios());
+        return "index";
     }
 }
