+++ src/main/java/com/login/app/modelo/Rol.java
@@ -1,8 +1,10 @@
 package com.login.app.modelo;
 
-import jakarta.persistence.*;
-
-import javax.annotation.processing.Generated;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 
 @Entity
 @Table(name="TiposUsuarios")
