+++ src/main/java/com/login/app/modelo/Usuario.java
@@ -1,9 +1,19 @@
 package com.login.app.modelo;
 
-import jakarta.persistence.*;
-import org.hibernate.sql.ast.tree.expression.Collation;
 import java.util.Collection;
 import java.util.List;
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.JoinTable;
+import javax.persistence.ManyToMany;
+import javax.persistence.Table;
+import javax.persistence.UniqueConstraint;
 
 @Entity
 @Table(name="usuarios", uniqueConstraints = @UniqueConstraint(columnNames = "email"))
