+++ src/main/java/com/login/app/servicio/UsuarioServicio.java
@@ -1,10 +1,13 @@
 package com.login.app.servicio;
 
-import com.login.app.dto.UsuarioRegistroDTO;
+import com.login.app.controlador.dto.UsuarioRegistroDTO;
 import com.login.app.modelo.Usuario;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
+
+import java.util.List;
 
 public  interface UsuarioServicio extends UserDetailsService {
     public Usuario guardarUsuario(UsuarioRegistroDTO registroDTO);
+
+    public List<Usuario> listarUsuarios();
 }
