+++ src/main/resources/templates/login.html
@@ -1,25 +1,35 @@
 <!DOCTYPE html>
-<html lang="en" xmlns:th = "http://thymeleaf.org">
+<html xmlns:th="http://www.thymeleaf.org">
 <head>
-    <meta charset="UTF-8">
+    <meta charset="utf-8">
     <title>Inicio de sesión</title>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css" integrity="sha384-rHyoN1iRsVXV4nD0JutlnGaslCJuC7uwjduW9SVrLvRYooPp2bWYgmgJQIXwl/Sp" crossorigin="anonymous">
+    <!-- Latest compiled and minified CSS -->
+    <link rel="stylesheet"
+          href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"
+          integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u"
+          crossorigin="anonymous">
+
+    <!-- Optional theme -->
+    <link rel="stylesheet"
+          href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css"
+          integrity="sha384-rHyoN1iRsVXV4nD0JutlnGaslCJuC7uwjduW9SVrLvRYooPp2bWYgmgJQIXwl/Sp"
+          crossorigin="anonymous">
 </head>
 <body>
+
 
 <nav class="navbar navbar-inverse navbar-fixed-top">
     <div class="container">
         <div class="navbar-header">
-            <button type="button" class="navbar-toogle collapsed"
-                    data-toogle = "collapse" data-targed ="#navbar" aria-expanded="false"
+            <button type="button" class="navbar-toggle collapsed"
+                    data-toggle="collapse" data-target="#navbar" aria-expanded="false"
                     aria-controls="navbar">
-                <span class="sr-only">Toogle navigation</span>
-                <span class="icon-bar"></span>
-                <span class="icon-bar"></span>
-                <span class="icon-bar"></span>
+                <span class="sr-only">Toggle navigation</span> <span
+                    class="icon-bar"></span> <span class="icon-bar"></span> <span
+                    class="icon-bar"></span>
             </button>
-            <a class="navbar-brand" href="#" th:href="@{/}">Registro e inicio de sesión</a>
+            <a class="navbar-brand" href="#" th:href="@{/}">Registro y
+                inicio de sesión</a>
         </div>
     </div>
 </nav>
@@ -27,20 +37,19 @@
 <br>
 <br>
 
-        <div class="container">
-            <div class="row">
-                <div class="col-md-6 col-md-offset-3">
-
+<div class="container">
+    <div class="row">
+        <div class="col-md-6 col-md-offset-3">
             <h1>Inicio de sesión</h1>
             <form th:action="@{/login}" method="post">
 
-                <div th:if=${param.error}>
+                <div th:if="${param.error}">
                     <div class="alert alert-danger">
                         Usuario o contraseña inválidos
                     </div>
                 </div>
 
-                <div th:if=${param.logout}>
+                <div th:if="${param.logout}">
                     <div class="alert alert-info">
                         Ha cerrado sesión exitosamente
                     </div>
@@ -53,32 +62,50 @@
 
                 <div class="form-group">
                     <label class="control-label" for="username">Nombre de usuario:</label>
-                    <input id="username" type="text" class="form-control" required autofocus="autofocus" placeholder="Digite su email">
+                    <input id="username" name="username" type="text" class="form-control" required autofocus="autofocus" placeholder="Digite su email">
                 </div>
 
                 <div class="form-group">
                     <label class="control-label" for="password">Contraseña:</label>
-                    <input id="password" type="password" class="form-control" required autofocus="autofocus" placeholder="Digite su contraseña">
+                    <input id="password" name="password" type="password" class="form-control" required autofocus="autofocus" placeholder="Digite su contraseña">
                 </div>
 
                 <div class="form-group">
-                   <div class="row">
+                    <div class="row">
                         <div class="col-md-6 col-md-offset-3">
                             <input type="submit" class="form-control btn btn-primary" id="login-submit" value="Iniciar sesión">
                         </div>
-                   </div>
+                    </div>
                 </div>
             </form>
-                    <div class="form-group">
-                        <span class="">
-                            Si eres un nuevo usuario, <a th:href="@{/registro}">regístrate aquí</a>
-                        </span>
-                    </div>
+            <div class="form-group">
+					<span>Si eres nuevo usuario <a th:href="@{/registro}">regístrate
+							aqui</a></span>
+            </div>
+
         </div>
     </div>
 </div>
 
 
-<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
+<!-- Latest compiled and minified JavaScript -->
+<script
+        src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"
+        integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa"
+        crossorigin="anonymous"></script>
 </body>
 </html>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
